Bug fixed. Change return type to long.

+++ src/org/yccheok/jstock/gui/JStockOptions.java
@@ -339,7 +339,7 @@ public class JStockOptions {
         this.newsVersion = newsVersion;
     }
 
-    public int getNewsVersion()
+    public long getNewsVersion()
     {
         return this.newsVersion;
     }
